backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cachetools==5.5.1
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.4.0
django-crispy-forms==2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ffmpeg-python==0.2.0
future==1.0.0
httplib2==0.22.0
idna==3.10
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
pyparsing==3.2.1
pysrt==1.1.2
python-dotenv==1.1.1
requests==2.32.3
rsa==4.9
sqlparse==0.4.4
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
yt-dlp==2025.6.30
gevent==24.5.0
gunicorn==23.0.0